requirements.txt
altgraph==0.17.4
packaging==24.1
pefile==2023.2.7
pillow==10.4.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pywin32-ctypes==0.2.2
setuptools==70.3.0
tkinterdnd2==0.3.0
pillow_heif==0.9.3
